+++ forms/forms.vcxproj
@@ -32,6 +32,7 @@
     <ClCompile Include="textbutton.cpp" />
     <ClCompile Include="textedit.cpp" />
     <ClCompile Include="ticker.cpp" />
+    <ClCompile Include="transactioncontrol.cpp" />
     <ClCompile Include="tristatebox.cpp" />
     <ClCompile Include="ui.cpp" />
   </ItemGroup>
@@ -51,6 +52,7 @@
     <ClInclude Include="textbutton.h" />
     <ClInclude Include="textedit.h" />
     <ClInclude Include="ticker.h" />
+    <ClInclude Include="transactioncontrol.h" />
     <ClInclude Include="tristatebox.h" />
     <ClInclude Include="ui.h" />
   </ItemGroup>
