+++ game/ui/base/aliencontainmentscreen.cpp
@@ -3,24 +3,18 @@
 #include "forms/graphic.h"
 #include "forms/graphicbutton.h"
 #include "forms/label.h"
-#include "forms/listbox.h"
 #include "forms/radiobutton.h"
 #include "forms/scrollbar.h"
+#include "forms/transactioncontrol.h"
 #include "forms/ui.h"
-#include "framework/configfile.h"
 #include "framework/data.h"
 #include "framework/event.h"
-#include "framework/font.h"
 #include "framework/framework.h"
-#include "framework/keycodes.h"
 #include "framework/logger.h"
 #include "framework/renderer.h"
 #include "game/state/city/base.h"
 #include "game/state/city/building.h"
-#include "game/state/city/vehicle.h"
 #include "game/state/gamestate.h"
-#include "game/state/rules/city/vammotype.h"
-#include "game/state/shared/organisation.h"
 #include "game/ui/base/basestage.h"
 #include "game/ui/general/messagebox.h"
 #include <array>
